remove old unneeded code
