+++ Nos/Views/Discover/DiscoverTab.swift
@@ -17,49 +17,13 @@ struct DiscoverTab: View {
     
     @State private var performingInitialLoad = true
     static let initialLoadTime = 2
-    @State private var relaySubscriptions = SubscriptionCancellables()
     @State private var isVisible = false
-    private var featuredAuthors: [String]
-    
+
     @StateObject private var searchController = SearchController()
     @State private var date = Date(timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.initialLoadTime))
 
     @State var predicate: NSPredicate = .false
 
-    // MARK: - Init
-    
-    init(featuredAuthors: [String] = Array(DiscoverTab.FeaturedAuthorCategory.all.npubs)) {
-        self.featuredAuthors = featuredAuthors
-    }
-
-    // MARK: - Internal
-    
-    func cancelSubscriptions() async {
-        relaySubscriptions.removeAll()
-    }
-    
-    func subscribeToNewEvents() async {
-        await cancelSubscriptions()
-        
-        let featuredFilter = Filter(
-            authorKeys: featuredAuthors.compactMap {
-                PublicKey(npub: $0)?.hex
-            },
-            kinds: [.text, .delete],
-            limit: 200
-        )
-
-        relaySubscriptions.append(await relayService.subscribeToEvents(matching: featuredFilter))
-    }
-
-    func updatePredicate() {
-        predicate = Event.extendedNetworkPredicate(
-            currentUser: currentUser,
-            featuredAuthors: featuredAuthors,
-            before: date
-        )
-    }
-    
     // MARK: - View
     
     var body: some View {
@@ -72,10 +36,8 @@ struct DiscoverTab: View {
                     )
                 } else {
                     DiscoverGrid(
-                        featuredAuthors: featuredAuthors,
-                        predicate: predicate,
-                        searchController: searchController,
-                        columns: $columns
+                        featuredAuthorCategory: .all,
+                        searchController: searchController
                     )
                 }
             }
@@ -101,18 +63,6 @@ struct DiscoverTab: View {
                 }
             }
             .animation(.easeInOut, value: columns)
-            .task { 
-                updatePredicate()
-            }
-            .refreshable {
-                date = .now
-            }
-            .onChange(of: date) {
-                updatePredicate()
-            }
-            .refreshable {
-                date = .now
-            }
             .onAppear {
                 if router.selectedTab == .discover {
                     isVisible = true
@@ -121,12 +71,9 @@ struct DiscoverTab: View {
             .onDisappear {
                 isVisible = false
             }
-            .onChange(of: isVisible) { 
+            .onChange(of: isVisible) {
                 if isVisible {
                     analytics.showedDiscover()
-                    Task { await subscribeToNewEvents() }
-                } else {
-                    Task { await cancelSubscriptions() }
                 }
             }
             .navigationDestination(for: Event.self) { note in
@@ -200,23 +147,19 @@ struct DiscoverTab_Previews: PreviewProvider {
     @State static var relayFilter: Relay?
     
     static var previews: some View {
-        if let publicKey = user.publicKey {
-            DiscoverTab(featuredAuthors: [publicKey.npub])
-                .environment(\.managedObjectContext, previewContext)
-                .environmentObject(relayService)
-                .environmentObject(router)
-                .environment(currentUser)
-                .onAppear { createTestData(in: previewContext) }
-
-            DiscoverTab(featuredAuthors: [publicKey.npub])
-                .environment(\.managedObjectContext, previewContext)
-                .environmentObject(relayService)
-                .environmentObject(router)
-                .environment(currentUser)
-                .onAppear { createTestData(in: previewContext) }
-                .previewDevice("iPad Air (5th generation)")
-        } else {
-            EmptyView()
-        }
+        DiscoverTab()
+            .environment(\.managedObjectContext, previewContext)
+            .environmentObject(relayService)
+            .environmentObject(router)
+            .environment(currentUser)
+            .onAppear { createTestData(in: previewContext) }
+        
+        DiscoverTab()
+            .environment(\.managedObjectContext, previewContext)
+            .environmentObject(relayService)
+            .environmentObject(router)
+            .environment(currentUser)
+            .onAppear { createTestData(in: previewContext) }
+            .previewDevice("iPad Air (5th generation)")
     }
 }
